+++ src/Services/Readers/CSV.php
@@ -6,16 +6,15 @@ use Box\Spout\Reader\Common\Creator\ReaderEntityFactory;
 use Box\Spout\Reader\CSV\Reader;
 use Box\Spout\Reader\CSV\RowIterator;
 use Box\Spout\Reader\CSV\Sheet;
-use Box\Spout\Reader\CSV\SheetIterator;
-use Exception;
-use LaravelEnso\DataImport\Exceptions\Import;
+use LaravelEnso\DataImport\Services\Readers\Reader as ReaderInterface;
 
-class CSV
+
+class CSV extends ReaderInterface
 {
-    private bool $open;
-    private Reader $reader;
+    protected bool $open;
+    protected Reader $reader;
 
-    public function __construct(private string $file)
+    public function __construct(protected string $file)
     {
         $this->open = false;
         $this->reader = ReaderEntityFactory::createCSVReader();
@@ -29,7 +28,7 @@ class CSV
         }
     }
 
-    public function rowIterator(string $sheet): RowIterator
+    public function rowIterator(): RowIterator
     {
         $iterator = $this->sheet()->getRowIterator();
         $iterator->rewind();
@@ -42,33 +41,5 @@ class CSV
         $iterator = $this->sheetIterator();
 
         return $iterator->current();
-    }
-
-    public function sheetIterator(): SheetIterator
-    {
-        $this->ensureIsOpen();
-
-        $iterator = $this->reader->getSheetIterator();
-        $iterator->rewind();
-
-        return $iterator;
-    }
-
-    private function ensureIsOpen(): void
-    {
-        if (! $this->open) {
-            $this->open();
-        }
-    }
-
-    private function open(): void
-    {
-        try {
-            $this->reader->open($this->file);
-        } catch (Exception) {
-            throw Import::fileNotReadable($this->file);
-        }
-
-        $this->open = true;
     }
 }
